+++ packages/core/src/dom/index.ts
@@ -7,7 +7,6 @@ export * from './hotkey/actions';
 export * from './hotkey/aria';
 export * from './hotkey/coordinator';
 export * from './hotkey/hotkey';
-export * from './media/to-media-host';
 export * from './media/types';
 export * from './store/features';
 export * from './store/selectors';
