+++ packages/core/tsdown.config.ts
@@ -12,6 +12,7 @@ const createConfig = (mode: BuildMode): UserConfig => ({
     dom: './src/dom/index.ts',
     'dom/media/dash/index': './src/dom/media/dash/index.ts',
     'dom/media/hls/index': './src/dom/media/hls/index.ts',
+    'dom/media/host/index': './src/dom/media/host/index.ts',
     'dom/media/custom-media-element/index': './src/dom/media/custom-media-element/index.ts',
     'dom/media/mux/index': './src/dom/media/mux/index.ts',
     'dom/media/native-hls/index': './src/dom/media/native-hls/index.ts',
